+++ config/settings.json.sample
@@ -1,0 +1,3 @@
+{
+
+}
