--- list_classes.py
@@ -1,9 +1,0 @@
-import json
-
-import lib.config as config
-
-with open(config.CLASSES_PATH, "r") as file:
-  classes = json.load(file)
-
-for cls in classes:
-  print(cls)
